fix: change population path to pointer

+++ include/evolution/es.hpp
@@ -5,5 +5,5 @@
 
 struct ES
 {
-    static void ES1(double toiu);
+    static void ES1(float toiu);
 };
